+++ src/components/InterfaceConfig/components/BgpNeighborModal/BgpNeighborModal.tsx
@@ -9,9 +9,10 @@ import {
   Message,
   Popup,
 } from "semantic-ui-react";
-import { useAuthToken } from "../../contexts/AuthTokenContext";
-import { formatISODate } from "../../utils/formatters";
-import { fetchBgpSettings, fetchBgpNeighbors } from "../../services/deviceApi";
+import { useAuthToken } from "../../../../contexts/AuthTokenContext";
+import { formatISODate } from "../../../../utils/formatters";
+import { fetchBgpSettings } from "../../api/settingsApi";
+import { fetchBgpNeighbors } from "../../api/gnmiApi";
 
 interface BgpNeighborModalProps {
   readonly hostname: string;
