+++ src/components/InterfaceConfig/components/ImportInterfaceModal.tsx
@@ -1,7 +1,7 @@
 import { Modal, Button } from "semantic-ui-react";
 import { useState } from "react";
 import { useNavigate } from "react-router";
-import { importInterfaces } from "../api/interfaceConfigApi";
+import { importInterfaces } from "../api/deviceApi";
 import { useAuthToken } from "../../../contexts/AuthTokenContext";
 
 type ImportInterfaceModalProps = {
